Bug fix in jphase generator, requiring a check when creating continuous PH random variables.
The constructors in DenseContPhaseVar now check that the vector is substochastic and that the matrix is a subgenerator. IIf this is not the case, an exception is thrown. This fixes observed problem in the generator. Also, the NeutsContPHGenerator now relies on SSJ random number generator to improve the random number generation.